fix(build): Zero-warning build + RK3399 hardware correction
Compiler warnings fixed:
- mm.c: Fix __heap_start/__heap_end extern declarations (char[] vs uintptr_t)
- printk.c: Mark unused 'left_align' variables
- bloodmoon.c: Cast tab_id to uint32_t for sign-compare warnings
- gui_widgets.c: Cast ternary branches to int32_t in draw_switch
- modem_eg25g.c: Add braces to while/if statements (-Wmisleading-indentation)
- sync.c + mutex.h: Change mutex->locked from int to uint32_t, add volatile
  to cas32() parameter to resolve qualifier mismatch
- syscall.c: Add missing 'false' argument to display_blit() call

Hardware correction (PinePhone Pro):
- Makefile: Add BOARD=pinephone-pro target with cortex-a72+crc+crypto
- board.c: Fix RK3399S -> RK3399
- rk3399.h: Complete RK3399 SoC register map (CRU, GRF, GPIO, I2C,
  UART, DSI, VOP, GICv3, RK808 PMIC, CPU OPP tables, pinmux)
- display_rk3399_dsi.h: Corrected 1440x720 panel dimensions,
  DSI/VOP/D-PHY init scaffold for RK3399

Builds clean with zero warnings. Kernel boots successfully through
all 8 init phases in QEMU virt.

+++ bloodmoon.c
@@ -192,7 +192,7 @@ int bm_tab_navigate(int tab_id, const char* url)
 {
     for (int i = 0; i < TAB_MAX; i++) {
         bm_tab_t* tab = &bm_tabs[i];
-        if (tab->in_use && tab->id == tab_id) {
+        if (tab->in_use && tab->id == (uint32_t)tab_id) {
             spin_lock(&tab->lock);
             strncpy(tab->url, url, TAB_URL_MAX - 1);
             tab->loading = 1;
@@ -227,7 +227,7 @@ void bm_tab_close(int tab_id)
 {
     for (int i = 0; i < TAB_MAX; i++) {
         bm_tab_t* tab = &bm_tabs[i];
-        if (tab->in_use && tab->id == tab_id) {
+        if (tab->in_use && tab->id == (uint32_t)tab_id) {
             spin_lock(&tab->lock);
             if (tab->content_buf) kfree(tab->content_buf);
             tab->in_use = 0;
@@ -245,7 +245,7 @@ void bm_tab_set_network(int tab_id, uint32_t network)
 {
     for (int i = 0; i < TAB_MAX; i++) {
         bm_tab_t* tab = &bm_tabs[i];
-        if (tab->in_use && tab->id == tab_id) {
+        if (tab->in_use && tab->id == (uint32_t)tab_id) {
             uint32_t old = tab->network;
             tab->network = network;
             printk(KERN_INFO "bloodmoon: tab %d: %s -> %s\n",
